Reorganize job models into per-model folders and top-level migrations package
models/ now has one folder per versioned object (job_state/, job_submission_metadata/) with v1.py files and current-version aliases in each folder's __init__. The migration machinery (registry, protocol schema) moves to syft_job/migrations/. Old syft_job.models.state / .config import paths are removed and all consumers updated.
