+++ packages/syft-job/src/syft_job/models/__init__.py
@@ -1,9 +1,5 @@
-from .current_job_state import JobState
-from .current_job_submission_metadata import JobSubmissionMetadata
-from .job_state_v1 import JobStateV1, JobStatus
-from .job_submission_metadata_v1 import JobSubmissionMetadataV1
-from .migration import job_registry
-from . import schema  # noqa: F401  (registers the current protocol schema on import)
+from .job_state import JobState, JobStateV1, JobStatus
+from .job_submission_metadata import JobSubmissionMetadata, JobSubmissionMetadataV1
 
 __all__ = [
     # Current-version aliases
@@ -13,6 +9,4 @@ __all__ = [
     # Versioned objects
     "JobSubmissionMetadataV1",
     "JobStateV1",
-    # Migration registry
-    "job_registry",
 ]
