--- packages/syft-job/src/syft_job/models/config.py
@@ -1,7 +1,0 @@
-from .current_job_submission_metadata import JobSubmissionMetadata
-from .job_submission_metadata_v1 import JobSubmissionMetadataV1
-
-__all__ = [
-    "JobSubmissionMetadata",
-    "JobSubmissionMetadataV1",
-]
