--- packages/syft-job/src/syft_job/models/current_job_state.py
@@ -1,4 +1,0 @@
-from .job_state_v1 import JobStateV1
-
-# The current version of the job state object.
-JobState = JobStateV1
