--- packages/syft-job/src/syft_job/models/current_job_submission_metadata.py
@@ -1,4 +1,0 @@
-from .job_submission_metadata_v1 import JobSubmissionMetadataV1
-
-# The current version of the job submission metadata object.
-JobSubmissionMetadata = JobSubmissionMetadataV1
