+++ packages/syft-job/src/syft_job/models/job_state/__init__.py
@@ -1,0 +1,10 @@
+from .v1 import JobStateV1, JobStatus
+
+# The current version of the job state object.
+JobState = JobStateV1
+
+__all__ = [
+    "JobState",
+    "JobStateV1",
+    "JobStatus",
+]
