+++ packages/syft-job/src/syft_job/models/job_submission_metadata/__init__.py
@@ -1,0 +1,9 @@
+from .v1 import JobSubmissionMetadataV1
+
+# The current version of the job submission metadata object.
+JobSubmissionMetadata = JobSubmissionMetadataV1
+
+__all__ = [
+    "JobSubmissionMetadata",
+    "JobSubmissionMetadataV1",
+]
