--- packages/syft-job/src/syft_job/models/state.py
@@ -1,8 +1,0 @@
-from .current_job_state import JobState
-from .job_state_v1 import JobStateV1, JobStatus
-
-__all__ = [
-    "JobState",
-    "JobStateV1",
-    "JobStatus",
-]
