+++ tests/unit/test_datasets_jobs_repr.py
@@ -141,8 +141,7 @@ def _make_job_info(name: str, status: str = "pending") -> JobInfo:
 
     from syft_job.client import JobClient
     from syft_job.config import SyftJobConfig
-    from syft_job.models.config import JobSubmissionMetadata
-    from syft_job.models.state import JobState, JobStatus
+    from syft_job.models import JobState, JobStatus, JobSubmissionMetadata
 
     config = SyftJobConfig(
         syftbox_folder=Path("/tmp/fake"), current_user_email="test@test.com"
